--- a/addons/iap_crm/models/crm_lead.py
+++ b/addons/iap_crm/models/crm_lead.py
@@ -7,7 +7,7 @@ from odoo import fields, models
 class Lead(models.Model):
     _inherit = 'crm.lead'
 
-    reveal_id = fields.Char(string='Reveal ID') # Technical ID of reveal request done by IAP
+    reveal_id = fields.Char(string='Reveal ID', index='btree_not_null') # Technical ID of reveal request done by IAP
 
     def _merge_get_fields(self):
         return super(Lead, self)._merge_get_fields() + ['reveal_id']