## Description
ir_cron_crm_reveal_lead (executed by OdooBot)
⤷ _process_lead_generation
  ⤷ _perform_reveal_service
    ⤷ _create_lead_from_response
      ⤷ `already_created_lead = self.env['crm.lead'].search([
     ('reveal_id', '=', result['clearbit_id'])])`
         ⤷ `Seq.Scan` on `crm.lead`

Also `search(...)` -> `search_count(..., limit=1)` as it's just an existence check, searching for all instances is not necessary.

## Benchmark
On a database with over 8-digits count of `crm.lead`:

|               | Before | After    |
|---------------|--------|----------|
| Timings (hot) | 6.4 s  | 0.041 ms |
| Buffers usage | 2.6 GB | 32KB     |

## Reference
task-3977976

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr