@robodoo 2024-12-18
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/191045.png)](https://mergebot.odoo.com/odoo/odoo/pull/191045)